Add unit and integration tests

+++ stats-manager/src/main/java/ru/verstache/gabella/model/stats/ServerStats.java
@@ -15,7 +15,7 @@ public class ServerStats {
     private LocalDateTime lastPlayedAt;
     private int totalPlayedMatches;
     private double averageMatchesPerDay;
-    private double maxMatchesPerDay;
+    private int maxMatchesPerDay;
     private LocalDate mostPlayedDay;
     private PlayerDto bestPlayer;
 }
